refactor(trips): consolidate trip-announcements into trips module
Merged trip-announcements/ into trips/ to eliminate the separate module boundary.
TripAnnouncementsController uses @Controller('v1/trips/:id') with no literal-segment
routes, so controller ordering has no impact; TripAnnouncementsController appended last.
Updated database/schema/index.ts, notifications.service.ts, and trips.service.ts
to point to the new schema path.
