+++ apps/api/src/app.module.ts
@@ -15,7 +15,6 @@ import { UsersModule } from '@/modules/users/users.module';
 import { AssetsModule } from '@/modules/assets/assets.module';
 import { CloudStorageModule } from '@/modules/cloud-storage/cloud-storage.module';
 import { GroupsModule } from '@/modules/groups/groups.module';
-import { TripAnnouncementsModule } from '@/modules/trip-announcements/trip-announcements.module';
 import { TripsModule } from '@/modules/trips/trips.module';
 import { UploadsModule } from '@/modules/uploads/uploads.module';
 import { I18nHelperModule } from '@/i18n/i18n.module';
@@ -34,7 +33,6 @@ import * as path from 'path';
     CloudStorageModule,
     GroupsModule,
     TripsModule,
-    TripAnnouncementsModule,
     UploadsModule,
     HealthModule,
     FeedbackModule,
