+++ apps/api/src/database/schema/index.ts
@@ -17,7 +17,7 @@ export * from '@/modules/groups/schema/groups.schema';
 export * from '@/modules/notifications/schema/notification-deliveries.schema';
 export * from '@/modules/notifications/schema/notifications.schema';
 export * from '@/modules/notifications/schema/user-fcm-tokens.schema';
-export * from '@/modules/trip-announcements/schema/trip-announcements.schema';
+export * from '@/modules/trips/schema/trip-announcements.schema';
 export * from '@/modules/trips/schema/group-trips.schema';
 export * from '@/modules/trips/schema/trip-destinations.schema';
 export * from '@/modules/trips/schema/trip-participants.schema';
