+++ apps/api/src/modules/trips/dto/trip-announcement-response.dto.ts
