+++ apps/api/src/modules/trips/schema/trip-announcements.schema.ts
