+++ apps/api/src/modules/trips/trip-announcements.controller.ts
