+++ apps/api/src/modules/trips/trip-announcements.service.ts
